refactor: centralize datetime utility and improve code clarity
- Create shared server/core/utils.py with get_utc_now() function
- Keep timezone-aware datetimes in JWT token creation (required for expiration)
- Use or_() from sqlalchemy for better query readability
- Add documentation explaining datetime handling patterns
- Update all files to import from shared utility module
