+++ server/models/post.py
@@ -1,6 +1,6 @@
 from typing import Optional
 from uuid import UUID, uuid4
-from datetime import datetime, timezone
+from datetime import datetime
 
 from sqlmodel import SQLModel, Field
 from typing import Optional, Any, Dict
@@ -8,9 +8,12 @@ from sqlalchemy import Column
 from sqlalchemy.dialects.postgresql import JSONB
 
 
-def _get_utc_now():
-    """Get current UTC datetime without timezone info (naive datetime)."""
-    return datetime.now(timezone.utc).replace(tzinfo=None)
+# Import get_utc_now from core utils at module level would create circular import
+# So we define a module-level function that can be used as default_factory
+def _default_created_at():
+    """Default factory for created_at field."""
+    from ..core.utils import get_utc_now
+    return get_utc_now()
 
 
 class Post(SQLModel, table=True):
@@ -26,5 +29,5 @@ class Post(SQLModel, table=True):
     content_json: Optional[Dict[str, Any]] = Field(default=None, sa_column=Column(JSONB))
     comments_count: int = 0
     likes_count: int = 0
-    created_at: datetime = Field(default_factory=_get_utc_now)
+    created_at: datetime = Field(default_factory=_default_created_at)
     updated_at: Optional[datetime] = None
